requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
EbookLib==0.18
Flask==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
gTTS==2.5.3
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
pydub==0.25.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
WTForms==3.1.2
